Refactor session authorities fetching
