+++ src/handlers/dkg/dkgMetaData/eventHandler.ts
@@ -6,7 +6,7 @@ import { createPublicKey } from "./publicKey"
 import { u16, Vec } from "@polkadot/types-codec"
 import { DkgRuntimePrimitivesCryptoPublic } from "@polkadot/types/lookup"
 import { ITuple } from "@polkadot/types-codec/types"
-import { createOrUpdateSession } from "../../session"
+import { createOrUpdateSession, fetchSessionAuthorizes } from "../../session"
 import { DKGAuthority } from "../../../types"
 import { AbstractInt } from "@polkadot/types-codec/abstract/Int"
 
@@ -72,83 +72,12 @@ export const dkgMetaDataEventHandler = async (event: SubstrateEvent) => {
       break
     case DKGMetaDataSection.MisbehaviourReportsSubmitted:
       {
-        const eventData = eventDecoded.as(
-          DKGMetaDataSection.MisbehaviourReportsSubmitted
-        )
-
         const blockNumber = eventDecoded.blockNumber
 
-        const authorities: Vec<DkgRuntimePrimitivesCryptoPublic> = (await api.query.dkg.authorities()) as any
-        const nextAuthorities: Vec<DkgRuntimePrimitivesCryptoPublic> = (await api.query.dkg.nextAuthorities()) as any
-
-        const bestAuthorities: Vec<ITuple<
-          [u16, DkgRuntimePrimitivesCryptoPublic]
-        >> = (await api.query.dkg.bestAuthorities()) as any
-        const nextBestAuthorities: Vec<ITuple<
-          [u16, DkgRuntimePrimitivesCryptoPublic]
-        >> = (await api.query.dkg.nextBestAuthorities()) as any
-
-        const authorityReputations = await api.query.dkg.authorityReputations.entries()
-        const currentAuthoritiesAccounts = await api.query.dkg.accountToAuthority.entries()
-        // Acount32 => auth Id
-        const authorityIdMap: Record<string, string> = {}
-        // auth Id => auth reputation
-        const authorityReputationMap: Record<string, string> = {}
-
-        currentAuthoritiesAccounts.forEach(([key, val]) => {
-          authorityIdMap[
-            key.args[0].toString().replace("0x", "")
-          ] = val.toString().replace("0x", "")
-        })
-        authorityReputations.forEach(([key, val]) => {
-          authorityReputationMap[
-            key.args[0].toString().replace("0x", "")
-          ] = val.toString()
-        })
-
-        const dkgAuthorityMapper = (
-          id: DkgRuntimePrimitivesCryptoPublic
-        ): DKGAuthority => {
-          const accountId = id.toString().replace("0x", "")
-          const authorityId = authorityIdMap[accountId]
-          return {
-            accountId,
-            reputation: authorityReputationMap[authorityId],
-            authorityId,
-          }
-        }
-        const dkgAuthorities: DKGAuthority[] = authorities.map(
-          dkgAuthorityMapper
-        )
-        const nextDkgAuthorities: DKGAuthority[] = nextAuthorities.map(
-          dkgAuthorityMapper
-        )
-        const bestDkgAuthorities: DKGAuthority[] = bestAuthorities
-          .sort((a, b) => {
-            const aOrder = (a[0] as AbstractInt).toNumber()
-            const bOrder = (b[0] as AbstractInt).toNumber()
-            return aOrder > bOrder ? 1 : aOrder < bOrder ? -1 : 0
-          })
-          .map(([_order, key]) => {
-            return dkgAuthorityMapper(key)
-          })
-
-        const nextBestDkgAuthorities: DKGAuthority[] = nextBestAuthorities
-          .sort((a, b) => {
-            const aOrder = (a[0] as AbstractInt).toNumber()
-            const bOrder = (b[0] as AbstractInt).toNumber()
-            return aOrder > bOrder ? 1 : aOrder < bOrder ? -1 : 0
-          })
-          .map(([_order, key]) => {
-            return dkgAuthorityMapper(key)
-          })
+        const sessionAuthorities = await fetchSessionAuthorizes(blockNumber)
 
         await createOrUpdateSession({
-          blockId: blockNumber,
-          authorities: dkgAuthorities,
-          nextAuthorities: nextDkgAuthorities,
-          bestAuthorities: bestDkgAuthorities,
-          nextBestAuthorities: nextBestDkgAuthorities,
+          ...sessionAuthorities,
         })
       }
       break
